ASoC: Intel: ssp-common: relocate header file
Moving ssp-common header file from sound/soc/intel/common directory to
include/sound directory and rename the file. Keep file content
unchanged for tracking purpose.

Link: https://msgid.link/r/20240327162408.63953-3-pierre-louis.bossart@linux.intel.com
